[PATCH] Fix SLP when ifcvt versioned loop is not vectorized
When ifcvt version a loop, it sets dont_vectorize to the scalar loop. If the
vector loop is not vectorized and removed, the scalar loop is still left with
dont_vectorize. As a result, BB vectorization will not happen.

This patch resets dont_vectorize to scalar loop when IFN_LOOP_VECTORIZED
is set to false.

gcc/ChangeLog:

	* tree-vectorizer.cc (pass_vectorize::execute): Reset dont_vectorize
	to scalar loop when setting IFN_LOOP_VECTORIZED to false.

gcc/testsuite/ChangeLog:

	* gcc.dg/vect/bb-slp-77.c: New test.