aarch64: remove falkor-tag-collision-avoidance pass
This code is not well tested and there is only a single testcase
(gcc.target/aarch64/pr94530.c) which only enables this code but it
is testing to make sure there is no ICE.

The falkor cores have not been supported from Qualcomm from 2019 or so
either. And I don't have a way to test these cores internally either.

Bootstrapped and tested on aarch64-linux-gnu.

gcc/ChangeLog:

	* config/aarch64/aarch64-passes.def: Don't add pass_tag_collision_avoidance.
	* config/aarch64/aarch64-protos.h (make_pass_tag_collision_avoidance): Remove.
	* config/aarch64/aarch64-tuning-flags.def (RENAME_LOAD_REGS): Remove.
	* config/aarch64/tuning_models/qdf24xx.h (qdf24xx_tunings): Set tuning flags to
	AARCH64_EXTRA_TUNE_NONE.
	* config/aarch64/falkor-tag-collision-avoidance.cc: Removed.
	* config/aarch64/t-aarch64 (falkor-tag-collision-avoidance.o): Remove.
	* config.gcc (aarch64*-*-*): Remove falkor-tag-collision-avoidance.o from extra_objs.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>