aarch64: Move ENTRY_VHSDF to aarch64-simd-pragma-builtins.def
It's more convenient for later patches if we only define ENTRY_VHSDF
once, in the .def file.  Then the only macro that needs to be defined
before including the file is ENTRY itself.

The patch also moves the architecture requirements out of the
individual ENTRY invocations into a block-level definition of
REQUIRED_EXTENSIONS.  This reduces cut-&-paste a little and makes
things more consistent with aarch64-sve-builtins*.def.

gcc/
	* config/aarch64/aarch64-builtins.cc (ENTRY): Remove the features
	argument and get the features from REQUIRED_EXTENSIONS instead.
	(ENTRY_VHSDF): Move definition to...
	* config/aarch64/aarch64-simd-pragma-builtins.def: ...here.
	Move the architecture requirements to REQUIRED_EXTENSIONS.