c++: Adjust docs and option descriptions for the publishing of C++23
Now that C++23 has been finally published, the following patch attempts
to mention it in the option descriptions and documentation.
Given that it has been published about 1.5 years after being finalized
and has the 14882:2024 document number pair rather than :2023, I wasn't
sure when exactly to use 2023 (as informal name) and when 2024 (as year
of publishing), so I've tried to use 2024 in standards.texi which talks
more formally about the standards and a note that it has been published
in 2024 when it is talked about more informally.
I remember at least one older edition has been published in January too,
but the ISO pages pretend it was published still in December of the previous
year, in this case it doesn't.

2024-11-01  Jakub Jelinek  <jakub@redhat.com>

gcc/
	* doc/standards.texi (C++ Language): Mention also the 2024
	revision and -std=gnu++23 option.
	* doc/invoke.texi (-std=): Adjust description of c++23, c++2b,
	gnu++23 and gnu++2b now that ISO C++ 14882:2024 is published.
gcc/c-family/
	* c.opt (std=c++2b, std=c++23, std=gnu++2b, std=gnu++23): Adjust
	description now that ISO C++ 14882:2024 is published.