[committed] Make LRA default for ft32 and remove -mlra option
I was looking to clean up an old patch I'm carrying in my tester.  My first
thought was that ft32 was likely going to be deprecated because it wasn't using
LRA -- which in turn would mean the patch in question could just be removed.

But then I checked, ft32 has an LRA option and if turned on it gets the exact
same test results as with reload.  While the port mentions a failure with
sieve.c, that's been there since the port was introduced in 2015.

It's working well enough that I think just converting it is the right thing to
do.  The testsuite patch which precipitated this one will follow separately.
I've kept the -mlra option for compatibility sake, but it's ignored.

Pushing to the trunk.

gcc/
	* config/ft32/ft32.cc (ft32_lra_p): Remove.
	(TARGET_LRA_P): Likewise.
	* config/ft32/ft32.opt: Make -mlra ignored.
	* doc/invoke.texi: Adjust documentation for -mlra on ft32.