c++, contracts: Ensure deferred-parsed contracts are handled for lambdas [PR117431].
Currently, we defer parsing of contracts until classes are definedm including
the closures for lambdas.  However the latter take a different path through
the parser and we have been failing to complete the parsing.

At code-gen time, we then ignore contracts with deferred parsed content which
leads to omitting the outlined checks (or an error for in-lined ones).

This fix checks for unparsed contracts once the lambda function body is complete.

	PR c++/117431

gcc/cp/ChangeLog:

	* parser.cc (cp_parser_lambda_body): Check for and handle deferred-
	parsed contracts.

gcc/testsuite/ChangeLog:

	* g++.dg/contracts/pr117431.C: New test.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>