aarch64: Fix incorrect LS64 documentation
As Yuta Mukai pointed out, the manual wrongly said that LS64 is
enabled by default for Armv8.7-A and above, and for Armv9.2-A
and above.  LS64 is not mandatory at any architecture level
(and the code correctly implemented that).

I think this was a leftover from an early version of the spec.

gcc/
	* doc/invoke.texi: Fix documentation of LS64 so that it's
	not implied by Armv8.7-A or Armv9.2-A.