arm: [MVE intrinsics] Check immediate is a multiple in a range
This patch adds support to check that an immediate is a multiple of a
given value in a given range.

This will be used for instance by scatter_base to check that offset is
in +/-4*[0..127].

Unlike require_immediate_range, require_immediate_range_multiple
accepts signed range bounds to handle the above case.

gcc/ChangeLog:

	* config/arm/arm-mve-builtins.cc (report_out_of_range_multiple):
	New.
	(function_checker::require_signed_immediate): New.
	(function_checker::require_immediate_range_multiple): New.
	* config/arm/arm-mve-builtins.h
	(function_checker::require_immediate_range_multiple): New.
	(function_checker::require_signed_immediate): New.