c++: allow array mem-init with -fpermissive [PR116634]
We've accidentally accepted this forever (at least as far back as 4.7), but
it's always been ill-formed; this was PR59465.  And we didn't accept it for
scalar types.  But rather than switch to a hard error for this code, let's
give a permerror so affected code can continue to work with -fpermissive.

	PR c++/116634

gcc/cp/ChangeLog:

	* init.cc (can_init_array_with_p): Allow PR59465 case with
	permerror.

gcc/testsuite/ChangeLog:

	* g++.dg/diagnostic/aggr-init1.C: Expect warning with -fpermissive.
	* g++.dg/init/array62.C: Adjust diagnostic.
	* g++.dg/init/array63.C: Adjust diagnostic.
	* g++.dg/init/array64.C: Adjust diagnostic.

(cherry picked from commit 3545aab00152ed3db1d7ce6ca4e1671dde276980)