testsuite: Fix up gcc.target/powerpc/safe-indirect-jump-3.c test [PR117444]
The test safe-indirect-jump-3.c FAILs on powerpc64le-linux with the change
in jump table generation behavior with commit r15-4756-g06bc3a734e8890,
since it is compiled without optimization and expects jump tables to be
generated.  Add an explicit -fjump-tables to dg-options to get the old
behavior back.

2024-11-05  Peter Bergner  <bergner@linux.ibm.com>

gcc/testsuite/
	PR testsuite/117444
	* gcc.target/powerpc/safe-indirect-jump-3.c: Add -fjump-tables to
	dg-options.