Add details output for assume processing.
The Assume pass simply produces results, with no indication of how it
arrived as the results it gets.  Add some output to the details listing.

The only functional change is when gori is used to calculate a range
more than once (ie, multiple uses), we now load the merged range rather
than just using the last calculated one.

	* tree-assume.cc (assume_query::assume_query): Add debug output.
	(assume_query::update_parms): Likewise.
	(assume_query::calculate_phi): Likewise.
	(assume_query::calculate_op): Likewise.  Also pick up any
	merged path values.
	(assume_query::calculate_stmt): Likewise.